+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Corretto stesso errore in sponsor.validator.ts nella funzione handleValidationErrors
   - Corretto stesso errore in staff.routes.ts nelle funzioni GET, PUT e DELETE
   - Corretto stesso errore in teamKits.routes.ts nelle funzioni GET, POST, PUT e DELETE
+  - Corretto stesso errore in organizationDocuments.routes.ts nelle funzioni POST, GET (download), PUT e DELETE
   - Modificati i return nelle funzioni per essere compatibili con TypeScript
 
 ## [2.1.0] - 2025-08-04